requirements.txt
requests
python-dotenv
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
fastapi
uvicorn
fastmcp